requirements.txt
beautifulsoup4==4.13.4
chromadb==1.0.15
dateparser==1.2.2
docx==0.2.4
ollama==0.5.1
pymupdf==1.26.3
python-dotenv==1.1.1
pytz==2025.2
Requests==2.32.4
